Bug fixes and feature requests

+++ apps/backend/src/bot/messages/eventChannelCalendar.message.ts
@@ -27,19 +27,27 @@ export async function buildEventChannelCalendarMessage(id: number, client: Disco
 			}
 		}
 	});
-	if (!channel) throw new Error(`Could not create event channel message, since an event channel with id '${id}' could not be found`);
+	if (!channel)
+		throw new Error(
+			`Could not create event channel message, since an event channel with id '${id}' could not be found`
+		);
 
-	const guild = channel.discordGuildId ? await $discord.getGuild(client, channel.discordGuildId) : await $discord.getSystemGuild(client);
+	const guild = channel.discordGuildId
+		? await $discord.getGuild(client, channel.discordGuildId)
+		: await $discord.getSystemGuild(client);
 	if (!guild) return null;
 
 	const calendarEmbed = new EmbedBuilder()
 		.setColor(PRIMARY_COLOR)
 		.setTitle(":page_with_curl: Upcoming Events")
-		.setDescription("Times are shown in your timezone" + (channel.events.length === 0 ? "\n\n*No scheduled events*" : ""))
+		.setDescription(
+			"Times are shown in your timezone" +
+				(channel.events.length === 0 ? "\n\n*No scheduled events*" : "")
+		)
 		.setFooter({
 			text: "Only events posted in this channel will be listed"
 		});
-	
+
 	const now = new Date();
 
 	const started: Event[] = [];
@@ -57,7 +65,9 @@ export async function buildEventChannelCalendarMessage(id: number, client: Disco
 		} else if (event.startAt.getTime() < now.getTime() + 7 * 24 * 60 * 60 * 1000) {
 			next7Days.push(event);
 		} else {
-			const monthString = new Intl.DateTimeFormat("en", { month: "long", year: "numeric" }).format(event.startAt).replace(new RegExp(`\\s*${now.getFullYear()}`), "");
+			const monthString = new Intl.DateTimeFormat("en", { month: "long", year: "numeric" })
+				.format(event.startAt)
+				.replace(new RegExp(`\\s*${now.getFullYear()}`), "");
 			if (!byMonth[monthString]) {
 				byMonth[monthString] = [];
 			}
@@ -66,56 +76,79 @@ export async function buildEventChannelCalendarMessage(id: number, client: Disco
 		}
 	}
 
-	const channelLink = `https://discord.com/channels/${channel.discordGuildId ?? guild.id}/${channel.discordId}/`;
+	const channelLink = `https://discord.com/channels/${channel.discordGuildId ?? guild.id}/${
+		channel.discordId
+	}/`;
 
 	if (started.length > 0) {
 		calendarEmbed.addFields({
 			name: ":green_circle: Started",
-			value: `>>> ${started.map(event => {
-				const startTime = Math.round(event.startAt!.getTime() / 1000);
-				return `<t:${startTime}:t> **[${event.name}](${channelLink + event.discordEventMessageId})** <t:${startTime}:R>`;
-			}).join("\n")}`
+			value: `>>> ${started
+				.map((event) => {
+					const startTime = Math.round(event.startAt!.getTime() / 1000);
+					return `<t:${startTime}:t> **[${event.name}](${
+						channelLink + event.discordEventMessageId
+					})** <t:${startTime}:R>`;
+				})
+				.join("\n")}`
 		});
 	}
 
 	if (next24Hours.length > 0) {
 		calendarEmbed.addFields({
 			name: "Next 24 hours",
-			value: `>>> ${next24Hours.map(event => {
-				const startTime = Math.round(event.startAt!.getTime() / 1000);
-				return `<t:${startTime}:t> **[${event.name}](${channelLink + event.discordEventMessageId})** <t:${startTime}:R>`;
-			}).join("\n")}`
+			value: `>>> ${next24Hours
+				.map((event) => {
+					const startTime = Math.round(event.startAt!.getTime() / 1000);
+					return `<t:${startTime}:t> **[${event.name}](${
+						channelLink + event.discordEventMessageId
+					})** <t:${startTime}:R>`;
+				})
+				.join("\n")}`
 		});
 	}
 
 	if (next7Days.length > 0) {
 		calendarEmbed.addFields({
 			name: "Next 7 days",
-			value: `>>> ${next7Days.map(event => {
-				const startTime = Math.round(event.startAt!.getTime() / 1000);
-				return `<t:${startTime}:F> **[${event.name}](${channelLink + event.discordEventMessageId})** <t:${startTime}:R>`;
-			}).join("\n")}`
+			value: `>>> ${next7Days
+				.map((event) => {
+					const startTime = Math.round(event.startAt!.getTime() / 1000);
+					return `<t:${startTime}:F> **[${event.name}](${
+						channelLink + event.discordEventMessageId
+					})** <t:${startTime}:R>`;
+				})
+				.join("\n")}`
 		});
 	}
 
 	for (const [month, events] of Object.entries(byMonth)) {
 		calendarEmbed.addFields({
 			name: month,
-			value: `>>> ${events.map(event => {
-				const startTime = Math.round(event.startAt!.getTime() / 1000);
-				return `<t:${startTime}:F> **[${event.name}](${channelLink + event.discordEventMessageId})** <t:${startTime}:R>`;
-			}).join("\n")}`
+			value: `>>> ${events
+				.map((event) => {
+					const startTime = Math.round(event.startAt!.getTime() / 1000);
+					return `<t:${startTime}:F> **[${event.name}](${
+						channelLink + event.discordEventMessageId
+					})** <t:${startTime}:R>`;
+				})
+				.join("\n")}`
 		});
 	}
 
 	return {
-		embeds: [calendarEmbed],
+		embeds: [calendarEmbed]
 	} satisfies BaseMessageOptions;
 }
 
 async function getDiscordChannel(client: DiscordClient, channel: EventChannel) {
-	const discordChannel = await $discord.getChannel(client, channel.discordId, channel.discordGuildId ?? undefined);
-	if (!discordChannel?.isTextBased() || !(discordChannel instanceof TextChannel)) throw new Error("Could not find event channel, or channel is somehow not text based");
+	const discordChannel = await $discord.getChannel(
+		client,
+		channel.discordId,
+		channel.discordGuildId ?? undefined
+	);
+	if (!discordChannel?.isTextBased() || !(discordChannel instanceof TextChannel))
+		throw new Error("Could not find event channel, or channel is somehow not text based");
 
 	return discordChannel;
 }
@@ -123,27 +156,34 @@ async function getDiscordChannel(client: DiscordClient, channel: EventChannel) {
 async function getEventChannelCalendarMessage(client: DiscordClient, channel: EventChannel) {
 	try {
 		const discordChannel = await getDiscordChannel(client, channel);
-		return channel.discordCalendarMessageId ? await discordChannel.messages.fetch(channel.discordCalendarMessageId) : null;
+		return channel.discordCalendarMessageId
+			? await discordChannel.messages.fetch(channel.discordCalendarMessageId)
+			: null;
 	} catch (err) {
 		return null;
 	}
 }
 
-export async function updateEventChannelCalendarMessage(client: DiscordClient, channel: EventChannel, repost = false) {
+export async function updateEventChannelCalendarMessage(
+	client: DiscordClient,
+	channel: EventChannel,
+	repost = false
+) {
 	try {
-		
 		const payload = await buildEventChannelCalendarMessage(channel.id, client);
 		if (!payload) throw new Error("Failed to build event message");
-		
+
 		const discordChannel = await getDiscordChannel(client, channel);
 		let message = await getEventChannelCalendarMessage(client, channel);
-		if (repost && message) {
+
+		if (repost && message && discordChannel.lastMessageId !== message.id) {
 			await message.delete();
+			message = null;
 		}
-		
-		if (repost || !message) {
+
+		if (!message) {
 			message = await discordChannel.send(payload);
-	
+
 			await database.eventChannel.update({
 				where: { id: channel.id },
 				data: {
@@ -154,6 +194,10 @@ export async function updateEventChannelCalendarMessage(client: DiscordClient, c
 			await message.edit(payload);
 		}
 	} catch (err) {
-		logger.error("Failed to update event channel calendar message", { channel: channel.id }, err);
+		logger.error(
+			"Failed to update event channel calendar message",
+			{ channel: channel.id },
+			err
+		);
 	}
 }
